services/auth-api/pyproject.toml
[project]
name = "auth-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    # ddtrace 4.x bundles all integrations — no separate extras needed
    "ddtrace>=4.7.1",
    "fastapi>=0.136.0",
    "uvicorn[standard]>=0.44.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "pydantic>=2.13.0",
]

[project.optional-dependencies]
test = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "httpx>=0.28.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "httpx>=0.28.1",
]
